Renames main component to App
